network warnings fixed

+++ source/Utilities/Network/xlet/xlet.h
@@ -75,13 +75,13 @@ namespace xlet {
 
 
 
-        DAWn::Events::Signal<uint64_t, std::thread::id> letIsListening;
-        DAWn::Events::Signal<uint64_t, std::string>     letOperationalError;
+        DAWn::Events::Signal<int32_t, std::thread::id>  letIsListening;
+        DAWn::Events::Signal<int32_t, std::string>      letOperationalError;
         DAWn::Events::Signal<>                          letInvalidSocketError;
         DAWn::Events::Signal<>                          letIsIINBOnly;
 
-        DAWn::Events::Signal<uint64_t, uint64_t >   letAcceptedANewConnection;
-        DAWn::Events::Signal<uint64_t>              letWillCloseConnection;
+        DAWn::Events::Signal<int32_t, int32_t>          letAcceptedANewConnection;
+        DAWn::Events::Signal<int32_t>                   letWillCloseConnection;
 
 
      protected:
